moved database python scripts to database package in dal package
