+++ BankingApp/DAL/BankAccountDAL/BankAccountDALImplementation.py
@@ -1,7 +1,7 @@
 import logging
 from typing import List
 from BankingApp.DAL.BankAccountDAL.BankAccountDALInterface import BankAccountDALInterface
-from BankingApp.DAL.config import Connect
+from BankingApp.DAL.Database.config import Connect
 from BankingApp.Entities.BankAccount import BankAccount
 from BankingApp.Entities.FailedTransaction import FailedTransaction
