+++ BankingApp/DAL/TransactionDAL/TransactionDALImplementation.py
@@ -1,6 +1,6 @@
 import logging
 from typing import List
-from BankingApp.DAL.config import Connect
+from BankingApp.DAL.Database.config import Connect
 from BankingApp.DAL.TransactionDAL.TransactionDALInterface import TransactionDALInterface
 from BankingApp.Entities.FailedTransaction import FailedTransaction
 from BankingApp.Entities.Transaction import Transaction
