+++ run.py
@@ -1,4 +1,4 @@
-from BankingApp.DAL import config
+from BankingApp.DAL.Database import config
 from BankingApp.FlaskAPI import create_back_end_api
 
 api = create_back_end_api(config)
